modifications function paths
