+++ src/lib/authFunctions.js
@@ -1,10 +1,10 @@
 /* eslint-disable max-len */
 import {
   getAuth, createUserWithEmailAndPassword, sendEmailVerification, signInWithEmailAndPassword,
-  signInWithPopup, GoogleAuthProvider, signOut,
+  signInWithPopup, GoogleAuthProvider, signOut, collection, addDoc, onSnapshot, deleteDoc, doc, updateDoc, getDoc, orderBy, query,
   // eslint-disable-next-line import/no-unresolved
-} from 'https://www.gstatic.com/firebasejs/9.6.11/firebase-auth.js';
-import { app } from './config.js';
+} from './firestoreFunctions.js';
+import { app, db } from './config.js';
 
 export const auth = getAuth(app);
 
@@ -29,3 +29,18 @@ export const signGoogle = () => signInWithPopup(auth, provider);
 
 // auth Log out
 export const signOutFirebase = () => signOut(auth);
+
+// Función para guardar post
+export const saveTask = (description, email, createdAt) => addDoc(collection(db, 'posts'), { description, email, createdAt });
+
+// Función para tener el  post en tiempo real y que se publique de forma descendente
+export const onSnapshotFb = (callback) => onSnapshot(query(collection(db, 'posts'), orderBy('createdAt', 'desc')), callback);
+
+// Función para borrar post
+export const deletePost = (id) => deleteDoc(doc(db, 'posts', id));
+
+// Función para editar post
+export const getPost = (id) => getDoc(doc(db, 'posts', id));
+export const editPost = (id, newPost) => {
+  updateDoc(doc(db, 'posts', id), newPost);
+};
